Refactoring

+++ run.sh
@@ -4,11 +4,12 @@ set -euo pipefail
 
 day=$1
 nr=$(printf "%02d" "$day")
-input=input$nr.txt
+input=${2:-inputs/input$nr.txt}
 
-if [[ ! -s "$input" ]]
+if [[ "$input" != "-" ]] && [[ ! -s "$input" ]]
 then
   echo "Download input for day $day"
+  mkdir -p "inputs"
   wget -q -O "$input" --header "Cookie: session=${SESSION:?is not set}" "https://adventofcode.com/2021/day/$day/input"
 fi
